[UPDATE] Kospi200 테이블 데이터 타입 최적화, scrape_kospi_top200 함수 가독성 개선

+++ laynet.Database/dbo/Tables/Kospi200.sql
@@ -1,9 +1,9 @@
 ﻿CREATE TABLE [dbo].[Kospi200] (
-    [Ranking]         INT           NOT NULL,
+    [Ranking]         TINYINT       NOT NULL,
     [CompanyName]     NVARCHAR (30) NOT NULL,
     [StockCode]       CHAR (6)      NOT NULL,
-    [MarketSum]       BIGINT        NOT NULL,
-    [AskingPriceUnit] INT           NOT NULL,
+    [MarketSum]       BIGINT		NOT NULL,
+    [AskingPriceUnit] SMALLINT      NOT NULL,
     CONSTRAINT [PK_Kospi200] PRIMARY KEY CLUSTERED ([Ranking] ASC, [StockCode] ASC)
 );
